Export name collision fix

+++ app/models/exports/submission_export.rb
@@ -10,6 +10,8 @@
 #
 # add_index :exports, [:term_id], name: :index_exports_on_term_id
 
+require 'zip'
+
 class Exports::SubmissionExport < Export
   prop_accessor :base_path, :solitary_path, :group_path, :extract_zips, :include_solitary_submissions, :include_group_submissions
 
@@ -29,11 +31,9 @@ class Exports::SubmissionExport < Export
   end
 
   def prepare_zip!(directory)
-    term.submissions.find_each do |submission|
-      submission.submission_assets.each do |submission_asset|
-        if should_add?(submission_asset) && File.exist?(submission_asset.file.to_s)
-          add_asset_to_zip(submission_asset, directory)
-        end
+    SubmissionAsset.for_term(term).includes(submission: :exercise).find_each do |submission_asset|
+      if should_add?(submission_asset) && File.exist?(submission_asset.file.to_s)
+        add_asset_to_zip(submission_asset, directory)
       end
     end
   end
@@ -61,28 +61,41 @@ class Exports::SubmissionExport < Export
     path = File.join(zip_tmp_dir, submission_asset_path(submission_asset))
 
     if extract_zips? && submission_asset.content_type == SubmissionAsset::Mime::ZIP
-      extraction_dir = File.dirname(path)
-      extract_zip_to(submission_asset.file.to_s, extraction_dir)
+      extraction_dir = path_without_extension(path)
+
+      begin
+        extract_zip_to(submission_asset.file.to_s, unique_path(extraction_dir))
+      rescue => e
+        # zip could not be extracted, falling back to hardlinking original file instead
+
+        FileUtils.rm_r(extraction_dir) if File.exist?(extraction_dir)
+        hardlink_file(submission_asset.file.to_s, unique_path(path))
+      end
     else
-      hardlink_file(submission_asset.file.to_s, path)
+      hardlink_file(submission_asset.file.to_s, unique_path(path))
     end
   end
 
   def extract_zip_to(zip_path, extraction_dir)
+    zip_name = File.basename(zip_path, File.extname(zip_path))
     Zip::File.open(zip_path) do |archive|
       archive.entries.each do |entry|
         # ignoring directories - as mkdir_p is more reliable
         next if entry.name[-1] == '/'
 
-        extraction_path = File.join(extraction_dir, entry.name)
+        # do not include zip file name twice
+        entry_name = entry.name
+        if entry_name.start_with?(zip_name)
+          entry_name = entry_name[zip_name.length..-1]
+        elsif entry_name.start_with?("/#{zip_name}")
+          entry_name = entry_name[(zip_name.length + 1)..-1]
+        end
+
+        extraction_path = File.join(extraction_dir, entry_name)
         FileUtils.mkdir_p(File.dirname(extraction_path))
         entry.extract(extraction_path)
       end
     end
-  rescue
-    # zip could not be extracted, hardlinking original file instead
-    FileUtils.rm_r(extraction_dir)
-    hardlink_file(zip_path, File.join(extraction_dir, File.basename(zip_path)))
   end
 
   def hardlink_file(source_path, dst_path)
@@ -119,6 +132,27 @@ class Exports::SubmissionExport < Export
     path % Hash[filled_placeholders]
   end
 
+  def unique_path(path)
+    return path unless File.exist? path
+
+    basepath = path_without_extension(path)
+    extname = File.extname(path)
+
+    i = 2
+    loop do
+      unique_path = "#{basepath}-#{i}#{extname}"
+      return unique_path unless File.exist?(unique_path)
+
+      i += 1
+    end
+  end
+
+  def path_without_extension(path)
+    ext = File.extname(path)
+
+    path[0..(-ext.length - 1)]
+  end
+
   def extract_placeholders(path)
     path.scan(/%\{([^\}]+)\}/).lazy.map(&:first).map(&:to_s).to_a
   end
@@ -146,7 +180,7 @@ class Exports::SubmissionExport < Export
     if term_registrations.any?
       grading_scale_service = GradingScaleService.new(term)
       grades = term_registrations.map { |tr| grading_scale_service.grade_for(tr) }
-      average = grades.reduce(:+).to_f / grades.length
+      average = grades.map(&:to_f).reduce(:+) / grades.length
 
       average.round.to_s
     else
